+++ src/components/UI/VideoCallStatus/VideoCallStatus.tsx
@@ -1,0 +1,26 @@
+import styles from './VideoCallStatus.module.scss'
+
+interface VideoCallStatusProps {
+  isConnected: boolean
+  isConnecting: boolean
+}
+
+export default function VideoCallStatus({
+  isConnected,
+  isConnecting,
+}: VideoCallStatusProps) {
+  return (
+    <div className={styles.connectionStatus}>
+      <strong
+        className={`${
+          isConnected
+            ? styles.connected
+            : isConnecting
+              ? styles.connecting
+              : styles.disconnected
+        }`}>
+        {isConnected ? ' ' : isConnecting ? 'Connecting...' : 'Disconnected'}
+      </strong>
+    </div>
+  )
+}